refactor: sync final graph hash and align directory to PaddleX
