--- graph_net/paddle_samples/vision/squeezenet1_1/graph_hash.txt
@@ -1,1 +1,0 @@
-e4f3a5c6f15919d2254bbb54f324e06f176251076e5685054924329e7d189b38
