--- graph_net/samples/paddle_samples/vision/squeezenet1_1/graph_hash.txt
@@ -1,1 +1,0 @@
-fe50e21a3d942119b84131723149ca314a6ca2dc77fdf4fde773542e4abd4675
