--- graph_net/samples/paddle_samples/vision/squeezenet1_1/input_meta.py
@@ -1,3 +1,0 @@
-import paddle
-
-input_meta = {"inputs": {"shape": [1, 3, 224, 224], "dtype": "float32"}}
