--- graph_net/samples/paddle_samples/vision/squeezenet1_1/weight_meta.py
@@ -1,3 +1,0 @@
-import paddle
-
-weight_meta = {}
