+++ graph_net/tests/test_squeezenet1_1_extract.py
@@ -2,7 +2,6 @@ import paddle
 import os
 from paddle.vision.models import squeezenet1_1
 
-
 def extract_squeezenet():
     os.environ["FLAGS_enable_pir_api"] = "1"
     paddle.enable_static()
@@ -14,10 +13,10 @@ def extract_squeezenet():
         model = squeezenet1_1(pretrained=False)
         model(inputs)
 
-    base_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-    # Updated save path to the new directory structure
+    # Correct path to the new root-level paddle_samples directory
+    base_dir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
     save_path = os.path.join(
-        base_dir, "paddle_samples/vision/squeezenet1_1/squeezenet1_1.json"
+        base_dir, "paddle_samples/PaddleX/squeezenet1_1/squeezenet1_1.json"
     )
 
     os.makedirs(os.path.dirname(save_path), exist_ok=True)
@@ -25,7 +24,6 @@ def extract_squeezenet():
     with open(save_path, "w") as f:
         f.write(str(main_program))
     print(f"Extract success! Saved to: {save_path}")
-
 
 if __name__ == "__main__":
     extract_squeezenet()
