+++ paddle_samples/PaddleX/squeezenet1_1/input_meta.py
