+++ paddle_samples/PaddleX/squeezenet1_1/squeezenet1_1.json
