+++ paddle_samples/PaddleX/squeezenet1_1/weight_meta.py
